CharSequenceFactory: simplify the single Text Range case
In fact, SingleTextRangeCharSequence is not needed at all. This class wrapped a
CharBuffer, but unlike the former (and LargeText and MultiTextRangeCharSequence
as well), a CharBuffer _does_ know how to address/build subsequences of itself.

Therefore, change the single TextRange case by returning the CharBuffer itself,
or a subsequence of it.