add support for custom charset decoder supplier
Default (REPORT) action of CharsetDecoder is not always desired. With new API
supplier of CharsetDecoder can be provided. This allows to customize
CharsetDecoder to user needs.

TextDecoder#nextRange was update to account for different CodingErrorActions.
endOfInput for decode method is now determined dynamically. This allows to
get access to UNDERFLOW CoderResult and avoid replacement of 1+ byte
character that happened to be at the upper bound of range.

NotThreadSafeLargeTextTest was updated to test all supported
CodingErrorActions.

Strings.repeat in tests were replaced with RandomStringUtils from
apache-commons lang3. RandomStringUtils generate string with characters of
various byte length (1-4 for UTF8) that is better for randomized testing
than fixed 1 byte string that was used before.

Also:

* updated guava to 30.1-android (guava versions 21+ require jdk8, android
  flavor is compatible with jdk7, see:
  https://github.com/google/guava/wiki/Compatibility#older-jdks)
* updated assertj-core to latest 2.x version (3.x is not compatible with jdk8)